+++ lib/prefab/errors/missing_env_var_error.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Prefab
+  module Errors
+    class MissingEnvVarError < Prefab::Error
+      def initialize(message)
+        super(message)
+      end
+    end
+  end
+end